requirements.txt
Streamlit==1.37.1
ultralytics==8.3.27
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
torch==2.5.0

